+++ R/utils-bib.R
@@ -27,7 +27,8 @@ get_bib_note <- function(x) {
 }
 
 guess_bibtype <- function(x) {
-  init_guess <- switch(tolower(x$type),
+  init_guess <- switch(
+    tolower(x$type),
     "article" = "article",
     "book" = "book",
     "manual" = "manual",